op-program: Refactor derivation task to be separate from bootstrapping the program and verifying outputs.
Makes the prepare, derive, verify steps more clearly separate.
